W254: the return-contract sweep — +2 MATCH, and the lever is now precisely bounded
**6610 -> 6612 first-party MATCH (80.3%).** 0 lost, 0 new over-window, 0 regressions.

Wave 253 discovered that correcting a `u64` return declaration to `u32` took `fcl_misc 003c9850` from
nd201 to 2 and was part of the four-change fix that MATCHed two `mdlFile` functions. A census found 83
NONMATCHING functions declared returning `u64`/`long`/`undefined8`, 44 at correct size. This wave tested
all of them.

## The lever is real but narrow, and the boundary is exact

**A `u64` return only costs bytes when the returned value is a COMPUTED 32-bit value needing
normalization. Returning a constant costs nothing at any width.**

W254Itf established this decisively: twelve `u64`-declared `itfPanel` functions, including all three at
nd1-nd5, produced a **byte-identical object** when corrected — because every one returns a constant
zero. I broadcast the refined filter mid-wave (read the retail epilogue first; proceed only if a
computed value is being normalized), and the remaining agents applied it to skip dozens of cosmetic
cases immediately.

**Where the return really is computed, it paid.** All the wins are allocator-style functions that return
a pointer via `move $v0,$sN`:

- `mdlFile 00326510` nd224 -> 200 (556 -> 552)
- `mdlFile 00332070` nd215 -> **187** (348 -> 344) — the return-only probe alone regressed to nd220, and
  correcting the associated allocator-result local to `u32` compensated the growth
- `mdlFile 003330e0` nd324 -> 319, `003393d0` nd197 -> 195
- `fclShopMisc 003fe3e0` nd206 -> 183; `003f39a0` nd689 -> 687
- `frTagTable 003af110` nd320 -> 311; `frFont 003b0620` nd279 -> 276 (also removing a stale `(long)`
  return cast); `y_misc 0042ac60` nd52 -> **33**

**And the family finished.** `mdlFile 0034b1a0` and `0034b6b0` both reached **nd0 (MATCH)** — the
two-instruction pre-JAL argument-setup-order residual that looked like a floor turned out to be
solvable. All six members of the `0034b*` family are now MATCH or nd0-adjacent, from a single
investigation begun two waves ago.

## Genuine 64-bit returns exist, and `daddiu` is the tell

My hypothesis that an `addiu`-versus-`daddiu` residual signals a width mismatch to be narrowed was
**wrong in the direction I guessed.** W254Btl found retail genuinely does 64-bit arithmetic in several
places: `btlBoss 002f88c0` feeds `$v0` with `daddiu`/`daddu` at six sites and the `u32` probe regressed
nd85 -> 212; `002fb860` returns through `dsll32`/`dsra32`; `fclShopMisc 003f0d60` returns the full saved
parameter via `daddu $v0,$s1` and uses `dsll32`/`dsra32` for signed-short cases. **In those functions
the wide declaration and the `long` locals are load-bearing and must stay.**

## A second negative: wide LOCALS are not the same lever

Narrowing `long`/`u64` locals regressed repeatedly — `itfPanel 003ad400` nd277 -> 326,
`y_runtime 0044aaf0` nd290 -> 295, `mdlFile 00337a60` nd150 -> 198, `00336920` nd782 -> 838. Locals
declared wide are frequently load-bearing for register-pair allocation. Test individually; expect
regressions.

## The safe caller pattern, exercised properly

`fclShopMisc 003f39a0` improved but direct narrowing **broke the matched sibling `003f3970`**. The agent
retained the original `u64` declaration as that caller's view and defined a `u32` aliased view for the
target, preserving all 158 matches. That is the pattern for this whole class: correct the definition,
keep a separate wider call view for any matched caller.

## Well-characterised residuals recorded for later

`btlCamera 002b21f0`'s missing retained float is `baseY` — retail keeps it in `$f21` across two calls
while ours reloads it; removing the pre-call `target.y` assignment reached nd890 but would pass an
uninitialized value, so it was correctly rejected. `btlCamera 002b8990`'s ascending-case structure
reproduces retail exactly at 440/448 but a four-byte register-colouring shift keeps nd at 255 versus the
194 baseline. `mt_evtMain 00369320` has an independent frame defect (ours retains `s2`, retail saves
s0-s1). `y_runtime 004312b0` (0x50/s0-s2 versus retail 0x40/s0-s1) and `00454620` (0xe0 versus 0xf0) are
live lifetime defects with the return contract now eliminated as a cause.

+++ src/frFont.c
@@ -153,7 +153,7 @@ u8 *FUN_003b0970(u8 *param_1,u8 param_2,u8 param_3,u8 param_4,u8 *param_5);
 #pragma alias FUN_003b0970_typed FUN_003b0970
 extern long FUN_003b0970_typed(void *param_1,u8 param_2,u8 param_3,u8 param_4,u32 param_5);
 void FUN_003b05c0(u16 *param_1);
-long FUN_003b0620(u16 param_1,u8 param_2,u8 param_3,u8 param_4);
+int FUN_003b0620(u16 param_1,u8 param_2,u8 param_3,u8 param_4);
 #pragma alias FUN_003b0620_typed FUN_003b0620
 extern int FUN_003b0620_typed(u16 param_1,u8 param_2,u8 param_3,u8 param_4);
 int FUN_003b0840(short *param_1);
@@ -1032,7 +1032,7 @@ void FUN_003b05c0(u16 *param_1)
 #undef FUN_003b0620
 // FUN_003B0620 NONMATCHING
 
-long FUN_003b0620(u16 param_1,u8 param_2,u8 param_3,u8 param_4)
+int FUN_003b0620(u16 param_1,u8 param_2,u8 param_3,u8 param_4)
 {
   u16 uVar5;
   int iVar6;
@@ -1089,7 +1089,7 @@ long FUN_003b0620(u16 param_1,u8 param_2,u8 param_3,u8 param_4)
       *(u32 *)(puVar7 + 8) = 0;
     }
   }
-  return (long)puVar7;
+  return (int)puVar7;
 }
 #define FUN_003b0620(...) ((long (*)(...))FUN_003b0620)(__VA_ARGS__)
 #undef FUN_003b0840
